Use cp to copy boot image for unencrypted disks.
Previously we always used spdk_dd to copy disk images to storage disk
files. For encrypted case there are no other easy ways, as data needs to
be encrypted in the disk file.

But for unencrypted disks, we could alternatively use Linux's `cp`
utility, which is faster. Based on an experiment I did, using `spdk_dd`
took 6.9 seconds to do an unencrypted copy of an image file, while `cp`
took about 3.2 seconds.

So this PR is changing the code to use `cp` for unencrypted disks.