Router Advertisement via radvd.
With this PR, we are not using Dnsmasq anymore for router advertisement
and moving the functionality out to radvd. The main reason to do so is
the limitation of dnsmasq for prefix delegation. In the next stages of
the project, we will start using Kea for DHCP since it supports prefix
delegation. Therefore, we need another tool to handle router
advertisement. Dnsmasq will someday be only used for DNS.